+++ cli/commands/up.go
@@ -2,34 +2,43 @@ package commands
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"os"
 	"os/signal"
 	"path/filepath"
-	"sync"
 	"syscall"
 
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/hashicorp/go-multierror"
+	"github.com/mattn/go-isatty"
+	"github.com/muesli/termenv"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/wundergraph/wundergraph/cli/helpers"
 	"github.com/wundergraph/wundergraph/pkg/bundler"
+	"github.com/wundergraph/wundergraph/pkg/cli"
 	"github.com/wundergraph/wundergraph/pkg/files"
+	"github.com/wundergraph/wundergraph/pkg/logging"
 	"github.com/wundergraph/wundergraph/pkg/node"
 	"github.com/wundergraph/wundergraph/pkg/operations"
 	"github.com/wundergraph/wundergraph/pkg/scriptrunner"
 	"github.com/wundergraph/wundergraph/pkg/telemetry"
 	"github.com/wundergraph/wundergraph/pkg/watcher"
 	"github.com/wundergraph/wundergraph/pkg/webhooks"
+	"github.com/wundergraph/wundergraph/pkg/wgpb"
 )
 
 const UpCmdName = "up"
 
 var (
-	upCmdPrettyLogging                      bool
 	defaultDataSourcePollingIntervalSeconds int
 	disableCache                            bool
 	clearCache                              bool
+	enableTUI                               bool
+	logs                                    bool
 )
 
 // upCmd represents the up command
@@ -41,6 +50,56 @@ var upCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx, cancel := context.WithCancel(context.Background())
 		defer cancel()
+
+		// Enable TUI only if stdout is a terminal
+		if !isatty.IsTerminal(os.Stdout.Fd()) {
+			// Always use JSON when not in a terminal
+			enableTUI = false
+		}
+
+		// Bubbletea UI is not compatible with regular stdout logging
+		if enableTUI {
+			if rootFlags.DebugMode {
+				log.Warn("Debug mode is enabled. This will disable the UI.")
+				enableTUI = false
+			} else if logs {
+				log.Warn("Logs are enabled. This will disable the UI.")
+				enableTUI = false
+			}
+		}
+
+		var devTUI *tea.Program
+		defaultOutput := termenv.DefaultOutput()
+
+		if enableTUI {
+
+			// For windows
+			restoreConsole, err := termenv.EnableVirtualTerminalProcessing(defaultOutput)
+			if err != nil {
+				return err
+			}
+			defer restoreConsole()
+
+			// Bubbletea UI is not compatible with regular stdout logging
+			// for those reasons we disable it and override the global logger. Any meaningful information
+			// should be logged to the UI or for deeper investigation the user should disable the UI and use the logging mode
+			log = zap.NewNop()
+
+			devTUI = cli.NewModel(ctx, &cli.Options{ServerVersion: BuildInfo.Version})
+
+			go func() {
+				if _, err := devTUI.Run(); err != nil {
+					log.Error("error running tui program", zap.Error(err))
+				}
+				cancel()
+			}()
+		}
+
+		reportErrToTUI := func(err error) {
+			if err != nil && devTUI != nil {
+				devTUI.Send(cli.Error{Err: err})
+			}
+		}
 
 		wunderGraphDir, err := files.FindWunderGraphDir(_wunderGraphDirConfig)
 		if err != nil {
@@ -61,7 +120,7 @@ var upCmd = &cobra.Command{
 			return err
 		}
 
-		// optional, no error check
+		// hook server is optional, so we don't error if it doesn't exist
 		codeServerFilePath, _ := files.CodeFilePath(wunderGraphDir, serverEntryPointFilename)
 
 		ctx, stop := signal.NotifyContext(ctx, os.Interrupt,
@@ -108,6 +167,7 @@ var upCmd = &cobra.Command{
 				WunderGraphDir: wunderGraphDir,
 				EnableCache:    !disableCache,
 			}),
+			Streaming: devTUI == nil,
 		})
 
 		// responsible for executing the config in "polling" mode
@@ -124,14 +184,54 @@ var upCmd = &cobra.Command{
 				EnableCache:                   !disableCache,
 				DefaultPollingIntervalSeconds: defaultDataSourcePollingIntervalSeconds,
 			}), "WG_DATA_SOURCE_POLLING_MODE=true"),
+			Streaming: devTUI == nil,
 		})
 
 		var hookServerRunner *scriptrunner.ScriptRunner
 		var webhooksBundler *bundler.Bundler
-		var onAfterBuild func() error
+		var onAfterBuild func(buildErr error, rebuild bool) error
 
 		outExtension := make(map[string]string)
 		outExtension[".js"] = ".cjs"
+
+		bundleOperations := func() error {
+			if files.DirectoryExists(operationsDir) {
+				operationsPaths, err := operations.GetPaths(wunderGraphDir)
+				if err != nil {
+					return err
+				}
+				err = operations.Cleanup(wunderGraphDir, operationsPaths)
+				if err != nil {
+					return err
+				}
+				err = operations.EnsureWunderGraphFactoryTS(wunderGraphDir)
+				if err != nil {
+					return err
+				}
+				operationsBundler := bundler.NewBundler(bundler.Config{
+					Name:          "operations-bundler",
+					EntryPoints:   operationsPaths,
+					AbsWorkingDir: wunderGraphDir,
+					OutDir:        generatedBundleOutDir,
+					OutExtension:  outExtension,
+					Logger:        log,
+				})
+				err = operationsBundler.Bundle()
+				if err != nil {
+					return err
+				}
+			}
+			return nil
+		}
+
+		onWatch := func(paths []string) {
+			if devTUI != nil {
+				task := cli.TaskStarted{
+					Label: "Building",
+				}
+				devTUI.Send(task)
+			}
+		}
 
 		if codeServerFilePath != "" {
 			hooksBundler := bundler.NewBundler(bundler.Config{
@@ -158,98 +258,112 @@ var upCmd = &cobra.Command{
 					OutDir:        generatedBundleOutDir,
 					OutExtension:  outExtension,
 					Logger:        log,
-					OnAfterBundle: func() error {
-						log.Debug("Webhooks bundled!", zap.String("bundlerName", "webhooks-bundler"))
+					OnAfterBundle: func(buildErr error, rebuild bool) error {
+						log.Debug("Webhooks bundled!", zap.String("bundlerName", "webhooks-bundler"), zap.Bool("rebuild", rebuild))
 						return nil
 					},
 				})
 			}
 
-			srvCfg := &helpers.ServerRunConfig{
+			srvCfg := &helpers.HooksServerRunConfig{
 				WunderGraphDirAbs: wunderGraphDir,
 				ServerScriptFile:  serverOutFile,
 				Env:               helpers.CliEnv(rootFlags),
 				Debug:             rootFlags.DebugMode,
+				LogStreaming:      devTUI == nil,
+				LogLevel:          rootFlags.CliLogLevel,
 			}
 
-			hookServerRunner = helpers.NewServerRunner(log, srvCfg)
+			hookServerRunner = helpers.NewHooksServerRunner(log, srvCfg)
 
-			onAfterBuild = func() error {
-				log.Debug("Config built!", zap.String("bundlerName", "config-bundler"))
+			onAfterBuild = func(buildErr error, rebuild bool) (err error) {
+				defer func() {
+					if devTUI != nil {
+						devTUI.Send(cli.TaskEnded{
+							Label: "Build completed",
+							Err:   err,
+						})
+					}
+				}()
 
-				if files.DirectoryExists(operationsDir) {
-					operationsPaths, err := operations.GetPaths(wunderGraphDir)
-					if err != nil {
-						return err
-					}
-					err = operations.Cleanup(wunderGraphDir, operationsPaths)
-					if err != nil {
-						return err
-					}
-					err = operations.EnsureWunderGraphFactoryTS(wunderGraphDir)
-					if err != nil {
-						return err
-					}
-					operationsBundler := bundler.NewBundler(bundler.Config{
-						Name:          "operations-bundler",
-						EntryPoints:   operationsPaths,
-						AbsWorkingDir: wunderGraphDir,
-						OutDir:        generatedBundleOutDir,
-						OutExtension:  outExtension,
-						Logger:        log,
-					})
-					err = operationsBundler.Bundle()
-					if err != nil {
-						return err
-					}
+				// if there is a build error, we don't want to continue
+				if buildErr != nil {
+					return buildErr
+				}
+
+				bundleOperationsErr := bundleOperations()
+				if bundleOperationsErr != nil {
+					return bundleOperationsErr
 				}
 
 				// generate new config
 				<-configRunner.Run(ctx)
-
-				var wg sync.WaitGroup
-
-				wg.Add(1)
-				go func() {
-					defer wg.Done()
-					// bundle hooks
-					_ = hooksBundler.Bundle()
-				}()
-
-				if webhooksBundler != nil {
-					wg.Add(1)
-					go func() {
-						defer wg.Done()
-						_ = webhooksBundler.Bundle()
-					}()
+				if configRunner.Error() != nil {
+					return configRunner.Error()
 				}
 
-				wg.Wait()
+				var wg errgroup.Group
+
+				wg.Go(func() error {
+					return hooksBundler.Bundle()
+				})
+
+				if webhooksBundler != nil {
+					wg.Go(func() error {
+						return webhooksBundler.Bundle()
+					})
+				}
+
+				if err := wg.Wait(); err != nil {
+					return err
+				}
 
 				go func() {
 					// run or restart hook server
 					<-hookServerRunner.Run(ctx)
+					reportErrToTUI(hookServerRunner.Error())
 				}()
 
 				go func() {
 					// run or restart the introspection poller
 					<-configIntrospectionRunner.Run(ctx)
+					reportErrToTUI(configIntrospectionRunner.Error())
 				}()
 
 				return nil
 			}
 		} else {
-			log.Info("hooks EntryPoint not found, skipping", zap.String("file", serverEntryPointFilename))
-			onAfterBuild = func() error {
+			log.Debug("hooks EntryPoint not found, skipping", zap.String("file", serverEntryPointFilename))
+
+			onAfterBuild = func(buildErr error, rebuild bool) (err error) {
+				defer func() {
+					devTUI.Send(cli.TaskEnded{
+						Label: "Build completed",
+						Err:   err,
+					})
+				}()
+
+				// if there is a build error, we don't want to continue
+				if buildErr != nil {
+					return buildErr
+				}
+
 				// generate new config
 				<-configRunner.Run(ctx)
+				if configRunner.Error() != nil {
+					return configRunner.Error()
+				}
 
 				go func() {
 					// run or restart the introspection poller
 					<-configIntrospectionRunner.Run(ctx)
+					reportErrToTUI(configIntrospectionRunner.Error())
 				}()
 
-				log.Debug("Config built!", zap.String("bundlerName", "config-bundler"))
+				bundleOperationsErr := bundleOperations()
+				if bundleOperationsErr != nil {
+					return bundleOperationsErr
+				}
 
 				return nil
 			}
@@ -281,8 +395,15 @@ var upCmd = &cobra.Command{
 			IgnorePaths: []string{
 				"node_modules",
 			},
+			OnWatch:       onWatch,
 			OnAfterBundle: onAfterBuild,
 		})
+
+		if devTUI != nil {
+			devTUI.Send(cli.TaskStarted{
+				Label: "Building",
+			})
+		}
 
 		err = configBundler.Bundle()
 		if err != nil {
@@ -316,19 +437,55 @@ var upCmd = &cobra.Command{
 			}
 		}()
 
-		n := node.New(ctx, BuildInfo, wunderGraphDir, log)
+		var nodeLogger *zap.Logger
+
+		if enableTUI {
+			nodeLogger = zap.NewNop()
+		} else {
+			nodeLogger = logging.New(rootFlags.PrettyLogs, rootFlags.DebugMode, zapLogLevel)
+		}
+
+		n := node.New(ctx, BuildInfo, wunderGraphDir, nodeLogger)
+
 		go func() {
 			configFile := filepath.Join(wunderGraphDir, "generated", "wundergraph.config.json")
-			err := n.StartBlocking(
+			options := []node.Option{
 				node.WithConfigFileChange(configFileChangeChan),
 				node.WithFileSystemConfig(configFile),
-				node.WithDebugMode(rootFlags.DebugMode),
 				node.WithInsecureCookies(),
 				node.WithIntrospection(true),
 				node.WithGitHubAuthDemo(GitHubAuthDemo),
-				node.WithPrettyLogging(rootFlags.PrettyLogs),
+				node.WithRequestLogging(rootFlags.DebugMode),
 				node.WithDevMode(),
-			)
+			}
+
+			if devTUI != nil {
+				options = append(options, node.WithServerConfigLoadHandler(func(config *node.WunderNodeConfig) {
+
+					// The file is guaranteed to exist, because the server is only started after the config was built
+					if data, err := os.ReadFile(filepath.Join(wunderGraphDir, "generated", "wundergraph.build_info.json")); err == nil {
+						var buildInfo wgpb.BuildInfo
+						if err := json.Unmarshal(data, &buildInfo); err == nil {
+							devTUI.Send(cli.ServerConfigLoaded{
+								Webhooks:                 buildInfo.Stats.TotalWebhooks,
+								Operations:               buildInfo.Stats.TotalOperations,
+								DatasourceConfigurations: buildInfo.Stats.TotalApis,
+								ServerURL:                "http://" + config.Api.PrimaryHost,
+								FileUploads:              buildInfo.Stats.HasUploadProvider,
+								Authentication:           buildInfo.Stats.HasAuthenticationProvider,
+								PlaygroundEnabled:        config.Api.EnableGraphqlEndpoint,
+								SdkVersion:               buildInfo.Sdk.Version,
+								WunderctlVersion:         buildInfo.Wunderctl.Version,
+							})
+						}
+					}
+				}))
+				options = append(options, node.WithServerErrorHandler(func(err error) {
+					reportErrToTUI(multierror.Append(errors.New("could not start server"), err))
+				}))
+			}
+
+			err := n.StartBlocking(options...)
 			if err != nil {
 				log.Error("node exited", zap.Error(err))
 				// exit context because we can't recover from a server start error
@@ -343,7 +500,7 @@ var upCmd = &cobra.Command{
 		// wait for context to be canceled (signal, context cancellation or via cancel())
 		<-ctx.Done()
 
-		log.Info("Context was canceled. Initialize WunderNode shutdown ....")
+		log.Debug("Context was canceled. Initialize WunderNode shutdown ....")
 
 		// close all listeners without waiting for them to finish
 		_ = n.Close()
@@ -355,10 +512,11 @@ var upCmd = &cobra.Command{
 }
 
 func init() {
-	upCmd.PersistentFlags().BoolVar(&upCmdPrettyLogging, "pretty-logging", true, "switches the logging to human readable format")
-	upCmd.PersistentFlags().IntVar(&defaultDataSourcePollingIntervalSeconds, "default-polling-interval", 5, "default polling interval for data sources")
-	upCmd.PersistentFlags().BoolVar(&disableCache, "no-cache", false, "disables local caches")
-	upCmd.PersistentFlags().BoolVar(&clearCache, "clear-cache", false, "clears local caches before startup")
+	upCmd.PersistentFlags().BoolVar(&enableTUI, "ui", true, "Enable terminal user interface")
+	upCmd.PersistentFlags().BoolVar(&logs, "logs", false, "Enable log mode. Useful for debugging")
+	upCmd.PersistentFlags().IntVar(&defaultDataSourcePollingIntervalSeconds, "default-polling-interval", 5, "Default polling interval for data sources")
+	upCmd.PersistentFlags().BoolVar(&disableCache, "no-cache", false, "Disables local caches")
+	upCmd.PersistentFlags().BoolVar(&clearCache, "clear-cache", false, "Clears local caches before startup")
 
 	rootCmd.AddCommand(upCmd)
 }