+++ docs-website/config/navigation.tsx
@@ -59,6 +59,14 @@ const navigation = [
 						title: 'Relay',
 						href: '/docs/getting-started/relay-quickstart',
 					},
+					{
+						title: 'SvelteKit',
+						href: '/docs/getting-started/sveltekit-quickstart',
+					},
+					{
+						title: 'Nuxt',
+						href: '/docs/getting-started/nuxt-quickstart',
+					},
 				],
 			},
 			{
@@ -98,6 +106,10 @@ const navigation = [
 						href: '/docs/examples/nextjs-relay',
 					},
 					{
+						title: 'SvelteKit + SSR with TanStack Query',
+						href: '/docs/examples/sveltekit',
+					},
+					{
 						title: 'Next.js + Clerk',
 						href: '/docs/examples/nextjs-clerk',
 					},
@@ -112,6 +124,10 @@ const navigation = [
 					{
 						title: 'Vite + React Relay',
 						href: '/docs/examples/vite-react-relay',
+					},
+					{
+						title: 'Vite + Svelte',
+						href: '/docs/examples/vite-svelte',
 					},
 					{
 						title: 'Hooks',
@@ -188,6 +204,10 @@ const navigation = [
 					{
 						title: 'Next.js Todos',
 						href: '/docs/examples/nextjs-todos',
+					},
+					{
+						title: 'Nuxt + Vue Query',
+						href: '/docs/examples/nuxt',
 					},
 				],
 			},
@@ -340,11 +360,6 @@ const navigation = [
 				title: 'Cookie-based Auth',
 				links: [
 					{
-						title: 'Overview',
-						navTitle: 'Cookie-based Auth',
-						href: '/docs/auth/cookie-based-auth',
-					},
-					{
 						title: 'OpenID Connect',
 						href: '/docs/auth/cookie-based-auth/openid-connect',
 					},
@@ -370,17 +385,16 @@ const navigation = [
 				title: 'Token-based Auth',
 				links: [
 					{
-						title: 'Overview',
-						navTitle: 'Token-based Auth',
-						href: '/docs/auth/token-based-auth',
+						title: 'OpenID Connect JWKS',
+						href: '/docs/auth/token-based-auth/openid-connect',
 					},
 					{
 						title: 'Auth.js (NextAuth)',
 						href: '/docs/auth/token-based-auth/auth-js',
 					},
 					{
-						title: 'OpenID Connect',
-						href: '/docs/auth/token-based-auth/openid-connect',
+						title: 'Clerk.com',
+						href: '/docs/auth/token-based-auth/clerk',
 					},
 				],
 			},
@@ -663,8 +677,16 @@ const navigation = [
 						href: '/docs/supported-frontend-frameworks/svelte',
 					},
 					{
+						title: 'SvelteKit',
+						href: '/docs/supported-frontend-frameworks/sveltekit',
+					},
+					{
 						title: 'Vue',
 						href: '/docs/supported-frontend-frameworks/vue',
+					},
+					{
+						title: 'Nuxt',
+						href: '/docs/supported-frontend-frameworks/nuxt',
 					},
 					{
 						title: 'SolidJS',
@@ -961,6 +983,10 @@ const navigation = [
 				href: '/docs/clients-reference/nextjs',
 			},
 			{
+				title: 'Svelte Query',
+				href: '/docs/clients-reference/svelte-query',
+			},
+			{
 				title: 'Next.js Relay',
 				href: '/docs/clients-reference/nextjs-relay',
 			},
@@ -1055,6 +1081,10 @@ const navigation = [
 					{
 						title: 'Token-based Authentication',
 						href: '/docs/wundergraph-config-ts-reference/configure-token-based-authentication',
+					},
+					{
+						title: 'Custom Claims',
+						href: '/docs/wundergraph-config-ts-reference/configure-custom-claims',
 					},
 					{
 						title: 'Authorization / RBAC',