+++ pkg/apihandler/apihandler.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/buger/jsonparser"
 	"github.com/cespare/xxhash"
+	"github.com/davecgh/go-spew/spew"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/securecookie"
 	"github.com/hashicorp/go-multierror"
@@ -102,11 +103,11 @@ type Builder struct {
 
 	cache apicache.Cache
 
-	insecureCookies     bool
-	forceHttpsRedirects bool
-	enableDebugMode     bool
-	enableIntrospection bool
-	devMode             bool
+	insecureCookies      bool
+	forceHttpsRedirects  bool
+	enableRequestLogging bool
+	enableIntrospection  bool
+	devMode              bool
 
 	renameTypeNames []resolve.RenameTypeName
 
@@ -117,7 +118,7 @@ type Builder struct {
 type BuilderConfig struct {
 	InsecureCookies            bool
 	ForceHttpsRedirects        bool
-	EnableDebugMode            bool
+	EnableRequestLogging       bool
 	EnableIntrospection        bool
 	GitHubAuthDemoClientID     string
 	GitHubAuthDemoClientSecret string
@@ -137,7 +138,7 @@ func NewBuilder(pool *pool.Pool,
 		insecureCookies:            config.InsecureCookies,
 		middlewareClient:           hooksClient,
 		forceHttpsRedirects:        config.ForceHttpsRedirects,
-		enableDebugMode:            config.EnableDebugMode,
+		enableRequestLogging:       config.EnableRequestLogging,
 		enableIntrospection:        config.EnableIntrospection,
 		githubAuthDemoClientID:     config.GitHubAuthDemoClientID,
 		githubAuthDemoClientSecret: config.GitHubAuthDemoClientSecret,
@@ -236,7 +237,7 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 		})
 	}
 
-	if r.enableDebugMode {
+	if r.enableRequestLogging {
 		r.router.Use(logRequestMiddleware(os.Stderr))
 	}
 
@@ -256,7 +257,7 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 	if err := r.registerAuth(r.insecureCookies); err != nil {
 		if !r.devMode {
 			// If authentication fails in production, consider this a fatal error
-			return nil, err
+			return streamClosers, err
 		}
 		r.log.Error("configuring auth", zap.Error(err))
 	}
@@ -316,6 +317,7 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 		err = r.registerOperation(operation)
 		if err != nil {
 			r.log.Error("registerOperation", zap.Error(err))
+			return streamClosers, err
 		}
 	}
 
@@ -1705,6 +1707,14 @@ func (h *SubscriptionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) 
 	requestLogger := h.log.With(logging.WithRequestIDFromContext(r.Context()))
 	r = setOperationMetaData(r, h.operation)
 
+	// recover from panic if one occured. Set err to nil otherwise.
+	defer func() {
+		if r := recover(); r != nil {
+			requestLogger.Error("panic recovered", zap.Any("panic", r))
+			w.WriteHeader(http.StatusInternalServerError)
+		}
+	}()
+
 	if proceed := h.rbacEnforcer.Enforce(r); !proceed {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
 		return
@@ -2433,6 +2443,14 @@ func setOperationMetaData(r *http.Request, operation *wgpb.Operation) *http.Requ
 }
 
 func getOperationMetaData(r *http.Request) *OperationMetaData {
+	spew.Dump("------------>\ngetOperationMetaData", r)
+	if r == nil {
+		return nil
+	}
+	ctx := r.Context()
+	if ctx == nil {
+		return nil
+	}
 	maybeMetaData := r.Context().Value("operationMetaData")
 	if maybeMetaData == nil {
 		return nil